profiling/cmd: split to files
