+++ profiling/cmd/catapult.go
@@ -1,80 +1,34 @@
+/*
+ *
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 package main
 
 import (
-	"context"
 	"encoding/binary"
-	"encoding/gob"
 	"encoding/json"
-	"flag"
 	"fmt"
-	"google.golang.org/grpc"
 	"google.golang.org/grpc/grpclog"
 	"google.golang.org/grpc/internal/profiling"
-	"google.golang.org/grpc/profiling/proto"
-	pspb "google.golang.org/grpc/profiling/proto/service"
-	"io"
 	"os"
 	"sort"
 	"strings"
 	"time"
 )
-
-var flagAddress = flag.String("address", "", "address of your remote target")
-var flagTimeout = flag.Int("timeout", 0, "network operations timeout in seconds to remote target (0 indicates unlimited)")
-
-var flagRetrieveSnapshot = flag.Bool("retrieve-snapshot", false, "connect to remote target and retrieve a profiling snapshot locally for processing")
-var flagSnapshot = flag.String("snapshot", "", "snapshot file path")
-
-var flagEnableProfiling = flag.Bool("enable-profiling", false, "enable profiling in remote target")
-var flagDisableProfiling = flag.Bool("disable-profiling", false, "disable profiling in remote target")
-
-var flagStreamStatsCatapultJson = flag.String("stream-stats-catapult-json", "", "transform a snapshot into catapult JSON")
-var flagStreamStatsFilter = flag.String("stream-stats-filter", "server,client", "comma-separated list of stat tags to filter for")
-
-func exactlyOneOf(opts ...bool) bool {
-	first := true
-	for _, o := range opts {
-		if !o {
-			continue
-		}
-
-		if first {
-			first = false
-		} else {
-			return false
-		}
-	}
-
-	return !first
-}
-
-func parseArgs() error {
-	flag.Parse()
-
-	if *flagAddress != "" {
-		if !exactlyOneOf(*flagEnableProfiling, *flagDisableProfiling, *flagRetrieveSnapshot) {
-			return fmt.Errorf("when -address is specified, you must include exactly only one of -enable-profiling, -disable-profiling, and -retrieve-snapshot")
-		}
-
-		if *flagStreamStatsCatapultJson != "" {
-			return fmt.Errorf("when -address is specified, you must not include -catapult-json")
-		}
-	} else {
-		if *flagEnableProfiling || *flagDisableProfiling || *flagRetrieveSnapshot {
-			return fmt.Errorf("when -address isn't specified, you must not include any of -enable-profiling, -disable-profiling, and -retrieve-snapshot")
-		}
-
-		if *flagStreamStatsCatapultJson == "" {
-			return fmt.Errorf("when -address isn't specified, you must include -catapult-json")
-		}
-	}
-
-	return nil
-}
-
-type snapshot struct {
-	StreamStats []*profiling.Stat
-}
 
 type jsonNode struct {
 	Name      string  `json:"name"`
@@ -329,22 +283,7 @@ func streamStatsCatapultJsonify(stat *profiling.Stat, base time.Time) []jsonNode
 	return result
 }
 
-func streamStatsCatapultJson(snapshotFileName string, streamStatsCatapultJsonFileName string) error {
-	grpclog.Infof("opening snapshot file %s", snapshotFileName)
-	snapshotFile, err := os.Open(snapshotFileName)
-	if err != nil {
-		grpclog.Errorf("cannot open %s: %v", snapshotFileName, err)
-		return err
-	}
-
-	grpclog.Infof("decoding snapshot file %s", snapshotFileName)
-	s := &snapshot{}
-	decoder := gob.NewDecoder(snapshotFile)
-	if err = decoder.Decode(s); err != nil {
-		grpclog.Errorf("cannot decode %s: %v", snapshotFileName, err)
-		return err
-	}
-
+func streamStatsCatapultJson(s *snapshot, streamStatsCatapultJsonFileName string) error {
 	grpclog.Infof("calculating stream stats filters")
 	filterArray := strings.Split(*flagStreamStatsFilter, ",")
 	filter := make(map[string]bool)
@@ -438,123 +377,4 @@ func streamStatsCatapultJson(snapshotFileName string, streamStatsCatapultJsonFil
 
 	grpclog.Infof("successfully wrote catapult JSON file %s", streamStatsCatapultJsonFileName)
 	return nil
-}
-
-func setEnabled(ctx context.Context, c pspb.ProfilingClient, enabled bool) error {
-	_, err := c.SetEnabled(ctx, &pspb.SetEnabledRequest{Enabled: enabled})
-	if err != nil {
-		grpclog.Printf("error calling SetEnabled: %v\n", err)
-		return err
-	}
-
-	grpclog.Printf("successfully set enabled = %v", enabled)
-	return nil
-}
-
-func retrieveSnapshot(ctx context.Context, c pspb.ProfilingClient, f string) error {
-	grpclog.Infof("establishing stream stats stream")
-	stream, err := c.GetStreamStats(ctx, &pspb.GetStreamStatsRequest{})
-	if err != nil {
-		grpclog.Errorf("error calling GetStreamStats: %v\n", err)
-		return err
-	}
-
-	s := &snapshot{StreamStats: make([]*profiling.Stat, 0)}
-
-	grpclog.Infof("receiving and processing stream stats")
-	for {
-		resp, err := stream.Recv()
-		if err != nil {
-			if err == io.EOF {
-				grpclog.Infof("received EOF, last message")
-				break
-			}
-			grpclog.Errorf("error recv: %v", err)
-			return err
-		}
-
-		stat := proto.StatProtoToStat(resp)
-		s.StreamStats = append(s.StreamStats, stat)
-	}
-
-	grpclog.Infof("creating snapshot file %s", f)
-	file, err := os.Create(f)
-	defer file.Close()
-	if err != nil {
-		grpclog.Errorf("cannot create %s: %v", f, err)
-		return err
-	}
-
-	grpclog.Infof("encoding data and writing to snapshot file %s", f)
-	encoder := gob.NewEncoder(file)
-	err = encoder.Encode(s)
-	if err != nil {
-		grpclog.Infof("error encoding: %v", err)
-		return err
-	}
-
-	grpclog.Infof("successfully wrote profiling snapshot to %s", f)
-	return nil
-}
-
-func remoteCommand() error {
-	ctx := context.Background()
-	if *flagTimeout > 0 {
-		ctx, _ = context.WithTimeout(context.Background(), time.Duration(*flagTimeout)*time.Second)
-	}
-
-	grpclog.Infof("dialing %s", *flagAddress)
-	cc, err := grpc.Dial(*flagAddress, grpc.WithInsecure())
-	if err != nil {
-		grpclog.Errorf("cannot dial %s: %v", *flagAddress, err)
-		return err
-	}
-	defer cc.Close()
-
-	c := pspb.NewProfilingClient(cc)
-
-	if *flagEnableProfiling || *flagDisableProfiling {
-		return setEnabled(ctx, c, *flagEnableProfiling)
-	} else if *flagRetrieveSnapshot {
-		return retrieveSnapshot(ctx, c, *flagSnapshot)
-	} else {
-		return fmt.Errorf("what should I do with the remote target?")
-	}
-}
-
-func localCommand() error {
-	if *flagSnapshot == "" {
-		return fmt.Errorf("-snapshot flag missing")
-	}
-
-	if *flagStreamStatsCatapultJson == "" {
-		return fmt.Errorf("what should I do with the snapshot file?")
-	}
-
-	if *flagStreamStatsCatapultJson != "" {
-		if err := streamStatsCatapultJson(*flagSnapshot, *flagStreamStatsCatapultJson); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func main() {
-	if err := parseArgs(); err != nil {
-		grpclog.Errorf("error parsing flags: %v", err)
-		os.Exit(1)
-	}
-
-	if *flagAddress != "" {
-		if err := remoteCommand(); err != nil {
-			grpclog.Errorf("error: %v", err)
-			os.Exit(1)
-		}
-	} else {
-		if err := localCommand(); err != nil {
-			grpclog.Errorf("error: %v", err)
-			os.Exit(1)
-		}
-	}
 }
