+++ profiling/cmd/local.go
@@ -1,0 +1,68 @@
+/*
+ *
+ * Copyright 2019 gRPC authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package main
+
+import (
+	"encoding/gob"
+	"fmt"
+	"google.golang.org/grpc/grpclog"
+	"os"
+)
+
+func loadSnapshot(snapshotFileName string) (*snapshot, error) {
+	grpclog.Infof("opening snapshot file %s", snapshotFileName)
+	snapshotFile, err := os.Open(snapshotFileName)
+	if err != nil {
+		grpclog.Errorf("cannot open %s: %v", snapshotFileName, err)
+		return nil, err
+	}
+
+	grpclog.Infof("decoding snapshot file %s", snapshotFileName)
+	s := &snapshot{}
+	decoder := gob.NewDecoder(snapshotFile)
+	if err = decoder.Decode(s); err != nil {
+		grpclog.Errorf("cannot decode %s: %v", snapshotFileName, err)
+		return nil, err
+	}
+
+	return s, nil
+}
+
+func localCommand() error {
+	if *flagSnapshot == "" {
+		return fmt.Errorf("-snapshot flag missing")
+	}
+
+	s, err := loadSnapshot(*flagSnapshot)
+	if err != nil {
+		return err
+	}
+
+	if *flagStreamStatsCatapultJson == "" {
+		return fmt.Errorf("what should I do with the snapshot file?")
+	}
+
+	if *flagStreamStatsCatapultJson != "" {
+		if err = streamStatsCatapultJson(s, *flagStreamStatsCatapultJson); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
